+++ src/java/lang/System.cpp
@@ -24,7 +24,7 @@ namespace bestofjava {
 
    /**
     * Returns the current time in milliseconds. Note that while the unit of time of the return value is a millisecond, the granularity of the value depends on the underlying operating system and may be larger. For example, many operating systems measure time in units of tens of milliseconds. */
-   long long System::currentTimeMillis() {
+   int64_t System::currentTimeMillis() {
       struct timeval tv;
       struct timezone tz;
       int res = gettimeofday(&tv, &tz);
